Refactoring + Confirm before deletion message

+++ tools/comment_functions.php
@@ -36,7 +36,7 @@ function displayVisitorComments() {
 
   try {
     $query = $conn->prepare(
-        "SELECT * FROM comments WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id IS NULL AND comment_author_id IS NULL"
+        "SELECT * FROM comments WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id IS NULL AND comment_author_id IS NULL ORDER BY comment_created DESC"
     );
     $query->execute();
   } catch (PDOException $e) {
@@ -71,7 +71,7 @@ function displayComments() {
 
   try {
     $query = $conn->prepare(
-        "SELECT * FROM comments, users WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id IS NULL AND users.user_id = comments.comment_author_id"
+        "SELECT * FROM comments, users WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id IS NULL AND users.user_id = comments.comment_author_id ORDER BY comment_created DESC"
     );
     $query->execute();
   } catch (PDOException $e) {
@@ -98,6 +98,7 @@ function displayComments() {
       echo displayVisitorNestedComment($row['comment_id']);
       echo "</div></div>";
   }
+  displayVisitorComments();
 }
 
 function displayNestedComment($target_id) {
@@ -106,7 +107,7 @@ function displayNestedComment($target_id) {
 
   
   try {
-    $query = $conn->prepare("SELECT * FROM comments, users WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id = $target_id AND users.user_id = comments.comment_author_id");
+    $query = $conn->prepare("SELECT * FROM comments, users WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id = $target_id AND users.user_id = comments.comment_author_id ORDER BY comment_created DESC");
     $query->execute();
   
 
@@ -137,7 +138,7 @@ function displayVisitorNestedComment($target_id) {
 
   
   try {
-    $query = $conn->prepare("SELECT * FROM comments WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id = $target_id AND comments.comment_author_id IS NULL");
+    $query = $conn->prepare("SELECT * FROM comments WHERE comment_post_id = $post_id AND comment_status_id = 4 AND comment_reply_id = $target_id AND comments.comment_author_id IS NULL ORDER BY comment_created DESC");
     $query->execute();
   
 
@@ -306,7 +307,7 @@ function declareComments() {
      <td class=\"td-div\"><a class=\"td-btn td-btn-view\" href='../index.php?source=blog_post&blog_id=" . $row['comment_post_id'] . "'>View</a></td>
      <td class=\"td-div\"><a class=\"td-btn td-btn-success\" href='index.php?source=view_all_comments&approve=" . $row['comment_id'] . "'>Approve</a></td>
      <td class=\"td-div\"><a class=\"td-btn td-btn-reject\" href='index.php?source=view_all_comments&reject=" . $row['comment_id'] . "'>Reject</a></td>
-     <td class=\"td-div\"><a class=\"td-btn td-btn-danger\" href='index.php?source=view_all_comments&delete=" . $row['comment_id'] . "'>Delete</a></td>
+     <td class=\"td-div\"><a onClick=\" javascript: return confirm('Are you sure that you wish to delete this item?')\" class=\"td-btn td-btn-danger\" href='index.php?source=view_all_comments&delete=" . $row['comment_id'] . "'>Delete</a></td>
      <td class=\"td-div\"><a class=\"td-btn td-btn-info\" href='index.php?source=edit_comments&edit=" . $row['comment_id'] . "'>Edit</a></td>
      </tr>";
   }
