+++ tools/user_functions.php
@@ -362,7 +362,7 @@ function declareUsers() {
         <td class=\"td-div\"><a class=\"td-btn td-btn-view\" href='../index.php?source=profile_page&page=" . $row['user_id'] . "'>View</a></td>
         <td class=\"td-div\"><a class=\"td-btn td-btn-success\" href='index.php?source=view_all_users&approve=" . $row['user_id'] . "'>Approve</a></td>
         <td class=\"td-div\"><a class=\"td-btn td-btn-reject\" href='index.php?source=view_all_users&reject=" . $row['user_id'] . "'>Reject</a></td>
-        <td class=\"td-div\"><a class=\"td-btn td-btn-danger\" href='index.php?source=view_all_users&delete=" . $row['user_id'] . "'>Delete</a></td>
+        <td class=\"td-div\"><a onClick=\" javascript: return confirm('Are you sure that you wish to delete this item?')\" class=\"td-btn td-btn-danger\" href='index.php?source=view_all_users&delete=" . $row['user_id'] . "'>Delete</a></td>
         <td class=\"td-div\"><a class=\"td-btn td-btn-info\" href='index.php?source=edit_user&edit=" . $row['user_id'] . "'>Edit</a></td>
      </tr>";
   }
